+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ### v0.3.8
 
 - Update dependencies
-- Added `Shikensu.Contrib.transformContent` as an alias for `renderContent`
+- Add `Shikensu.Contrib.transformContent` as an alias for `renderContent`
+- Add `Shikensu.Contrib.setContent`
 
 ### v0.3.7